+++ mordant/src/commonMain/kotlin/com/github/ajalt/mordant/terminal/TerminalCursor.kt
@@ -1,7 +1,6 @@
 package com.github.ajalt.mordant.terminal
 
 import com.github.ajalt.mordant.internal.CSI
-import com.github.ajalt.mordant.widgets.RawWidget
 
 interface TerminalCursor {
     /**