+++ mordant/src/commonTest/kotlin/com/github/ajalt/mordant/animation/AnimationTest.kt
@@ -13,6 +13,34 @@ class AnimationTest {
     private val t = Terminal(terminalInterface = rec)
 
     @Test
+    @JsName("no_trailing_linebreak")
+    fun `no trailing linebreak`() {
+        val a = t.textAnimation<Int>(trailingLinebreak = false) { "<$it>\n===" }
+        a.update(1)
+        rec.output() shouldBe "<1>\n==="
+
+        // update
+        rec.clearOutput()
+        a.update(2)
+        val moves = t.cursor.getMoves { startOfLine(); up(1) }
+        rec.output() shouldBe "${moves}<2>\n==="
+    }
+
+    @Test
+    @JsName("no_trailing_linebreak_single_line")
+    fun `no trailing linebreak single line`() {
+        val a = t.textAnimation<Int>(trailingLinebreak = false) { "<$it>" }
+        a.update(1)
+        rec.output() shouldBe "<1>"
+
+        // update
+        rec.clearOutput()
+        a.update(2)
+        val moves = t.cursor.getMoves { startOfLine() }
+        rec.output() shouldBe "${moves}<2>"
+    }
+
+    @Test
     @JsName("print_during_animation")
     fun `print during animation`() {
         val a = t.textAnimation<Int> { "<$it>\n===" }